pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "netbox-yc-sync"
version = "3.0.0"
description = "Sync Yandex Cloud infrastructure to NetBox"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "vlkuz-dev"},
]
keywords = ["netbox", "yandex-cloud", "sync", "cmdb", "infrastructure"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "httpx>=0.25.0",
    "pynetbox>=7.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.4.0",
]

[project.urls]
Homepage = "https://github.com/vlkuz-dev/netbox-yc-sync"
Repository = "https://github.com/vlkuz-dev/netbox-yc-sync"
Issues = "https://github.com/vlkuz-dev/netbox-yc-sync/issues"

[project.scripts]
netbox-sync = "netbox_sync.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 120
